backend: extract peer-share logic into peer_share.go
Pulls the peer.Client wiring out of backend/radiance.go (was 1262 lines)
into a new same-package file so the dispatch / lifecycle / IPC accessor
all live together rather than scattered across 80+ lines of the main
file.

Moved to backend/peer_share.go:
  - peerController interface (the test-friendly subset of *peer.Client)
  - peerToggleTimeout constant
  - newPeerClient: peer.Client construction helper (used by NewLocalBackend)
  - applyPeerShare method
  - resumePeerShareIfEnabled method
  - closePeerClient method (used by Close; absorbs the nil-vpnClient
    guard pattern and the in-flight-resume Wait that were inline before)
  - PeerStatus method

backend/radiance.go now holds only the struct-field declarations and
six call sites (NewLocalBackend, Start, Close, PatchSettings dispatch).
No "github.com/getlantern/radiance/peer" import in radiance.go anymore.

Net diff: -84 lines in radiance.go (1262 → 1181), +121 lines in
peer_share.go. Tests pass.
